fix thread_usage.h
