+++ opal/mca/threads/base/Makefile.am
@@ -19,8 +19,7 @@
 #
 
 headers += \
-        base/base.h \
-        base/threads_base_null.h
+        base/base.h
 
 libmca_threads_la_SOURCES += \
         base/threads_base_open.c
